Fix integer indexing when reading intan files

+++ intan.py
@@ -273,8 +273,8 @@ class IntanFile(object):
         self.fid.seek(start_pos, 0)
 
         # Timestamps + DigIn + DigOut + ADCs
-        size_of_one = 4 + 2 + 2 + 2 * self.header["NumADCs"]
-        length = (file_size-start_pos) / size_of_one
+        size_of_one = int(4 + 2 + 2 + 2 * self.header["NumADCs"])
+        length = int((file_size-start_pos) / size_of_one)
 
         # Read in the whole thing at once
         byte_array = np.fromfile(self.fid, np.uint8, count=length*size_of_one)
